# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-tenant SaaS POS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List all organizations (super admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"

  # Branches
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders/transactions
      parameters:
        - { name: branchId, in: query, schema: { type: integer } }
        - { name: customerId, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order (void, hold, etc.)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/complete:
    post:
      operationId: completeOrder
      tags: [orders]
      summary: Complete order with payment
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/refund:
    post:
      operationId: refundOrder
      tags: [orders]
      summary: Refund an order
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Suppliers
...
components:
  schemas:
...
    Organization:
      type: object
      required: [id, name, businessType, plan, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        businessType: { type: string }
        logo: { type: ["string", "null"] }
        plan: { type: string, enum: [free, starter, professional, enterprise] }
        status: { type: string, enum: [active, suspended, trial] }
        currency: { type: string }
        timezone: { type: string }
        taxRate: { type: number }
        receiptHeader: { type: ["string", "null"] }
        receiptFooter: { type: ["string", "null"] }
        branchCount: { type: integer }
        userCount: { type: integer }
        createdAt: { type: string, format: date-time }
    OrganizationInput:
      type: object
      required: [name, businessType]
      properties:
        name: { type: string }
        businessType: { type: string }
        currency: { type: string }
        timezone: { type: string }
        taxRate: { type: number }
        plan: { type: string }
    OrganizationUpdate:
      type: object
      properties:
        name: { type: string }
        businessType: { type: string }
        currency: { type: string }
        timezone: { type: string }
        taxRate: { type: number }
        receiptHeader: { type: string }
        receiptFooter: { type: string }
        plan: { type: string }
        status: { type: string }

    # Branch
...
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        discount: { type: number }
        totalPrice: { type: number }
        notes: { type: ["string", "null"] }
    Order:
      type: object
      required: [id, orderNumber, branchId, status, subtotal, total, createdAt]
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        branchId: { type: integer }
        branchName: { type: string }
        customerId: { type: ["integer", "null"] }
        customerName: { type: ["string", "null"] }
        cashierId: { type: integer }
        cashierName: { type: string }
        status: { type: string, enum: [draft, held, completed, refunded, voided] }
        items: { type: array, items: { $ref: "#/components/schemas/OrderItem" } }
        subtotal: { type: number }
        discountAmount: { type: number }
        taxAmount: { type: number }
        total: { type: number }
        paymentMethod: { type: ["string", "null"] }
        amountPaid: { type: ["number", "null"] }
        change: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"] }
    OrderItemInput:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        unitPrice: { type: number }
        discount: { type: number }
        notes: { type: string }
    OrderInput:
      type: object
      required: [branchId, items]
      properties:
        branchId: { type: integer }
        customerId: { type: integer }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        discountAmount: { type: number }
        notes: { type: string }
    OrderUpdate:
      type: object
      properties:
        status: { type: string }
        notes: { type: string }
        customerId: { type: integer }
        discountAmount: { type: number }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    PaymentInput:
      type: object
      required: [paymentMethod, amountPaid]
      properties:
        paymentMethod: { type: string, enum: [cash, card, bank_transfer, mobile_wallet, qr, store_credit, mixed] }
        amountPaid: { type: number }
    RefundInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
        amount: { type: number }

    # Supplier
...