# lib/api-zod/src/generated/types/organizationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant SaaS POS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OrganizationInput {
  name: string;
  businessType: string;
  currency?: string;
  timezone?: string;
  taxRate?: number;
  plan?: string;
}